+++ Makefile
@@ -1,9 +1,8 @@
-DUMMY: test format
+DUMMY: test format path
 
 # Run the tests with `make test` command
 test:
-		pytest countant_test.py && pytest tests
+		pytest tests
 
 format:
 		black .
-
